day10 - 2023
Cool day, had fun creating the Pipeline class and found the solution with a little bit of debugging. The output of get of findConnection() made the second part of the problem easily solvable, it took some time to find out which cells were inside and outside and the solution I found isn't great but it works!